Update kafka-jars to v0.8.1.1.pre0
I decided to push a prerelease of the kafka-jars gem because it uses
zookeeper-jars and slf4j-jars with different versions from what is
actually provided in the binary release(s) of kafka.

I also removed usage of `bundle show <GEM>` in favour of using GEM_HOME
as it took forever to start kafka (due to starting new JVMs for each
`bundle` call).