Make ZK hosts first arg. to ZookeeperConsumer
Producer and Consumer already takes “hosts” as the first argument,
though in that case it’s actually the addresses to the brokers, but I
feel like it would be nice if all producer / consumer implementations
have similar constructors.

Note that I’m not certain if this is the most appropriate way, as in
Ruby I suppose it’s more common to just take an `options` hash, but oh
well.

Perhaps all constructors should default to localhost:<port> as well…?