Depend on `v0.8.1.1.pre1` release of kafka-jars
Just pushed a new pre-release version that depends on some other `jar`
gems instead of a massive jar with everything included. I decided to
keep it as a pre-release version for the time being, and when I’ve
actually tested it on a live Kafka cluster and everything seems to be
working fine, I’ll release the actual `v0.8.1.1` version.